stmgr: Split upgrade code from upgrade runtime
